remove duplicate Diagtician implementation
