+++ dlrover/python/master/diagnosis/diagnosis.py
@@ -13,51 +13,84 @@
 
 from typing import List
 
-from diagnosis.inferencechain.inferenceoperator.resolver.resolve_training_hang_operator import (  # noqa: E501
-    ResolveTrainingHangOperator,
+from dlrover.python.common.log import default_logger as logger
+from dlrover.python.diagnosis.common.diagnosis_action import DiagnosisAction
+from dlrover.python.diagnosis.inferencechain.coordinator import (
+    coordinate_solutions,
 )
-
 from dlrover.python.diagnosis.inferencechain.inference_chain import (
     Inference,
     InferenceChain,
     InferenceOperator,
-)
-from dlrover.python.diagnosis.inferencechain.inferenceoperator.observer.check_training_hang_operator import (  # noqa: E501
-    CheckTrainingHangOperator,
+    combine_inferences,
 )
 
 
 class Diagnostician:
+    """
+    Diagnostician is to observe training problems and explore solutions to
+    those problems during training.
+    """
+
     def __init__(self, data_manager):
         self._data_manager = data_manager
-        self._pre_checks: List[Inference] = []
         self._training_problems: List[Inference] = []
-
-    def get_pre_check_operators(self) -> List[InferenceOperator]:
-        return []
+        self._observers: List[InferenceOperator] = []
+        self._resolvers: List[InferenceOperator] = []
 
     def get_observing_operators(self) -> List[InferenceOperator]:
-        return [CheckTrainingHangOperator(self._data_manager)]
+        return self._observers
 
     def get_resolving_operators(self) -> List[InferenceOperator]:
-        return [ResolveTrainingHangOperator(self._data_manager)]
+        return self._resolvers
 
-    def register_pre_check(self, pre_checks: List[Inference]):
-        self._pre_checks = pre_checks
-
-    def register_problems(self, problems: List[Inference]):
+    def register_training_problems(self, problems: List[Inference]):
         self._training_problems = problems
 
-    def check_training(self) -> List[Inference]:
-        ic = InferenceChain(self._pre_checks, self.get_pre_check_operators())
-        return ic.infer()
+    def register_observers(self, operators: List[InferenceOperator]):
+        self._observers = operators
+
+    def register_resolvers(self, operators: List[InferenceOperator]):
+        self._resolvers = operators
 
     def observe_training(self) -> List[Inference]:
-        ic = InferenceChain(
-            self._training_problems, self.get_observing_operators()
-        )
-        return ic.infer()
+        """
+        To check if any problem in _training_problems happen
 
-    def diagnose_failure(self, observed_result: Inference) -> List[Inference]:
-        ic = InferenceChain([observed_result], self.get_resolving_operators())
-        return ic.infer()
+        Return:
+            observed problems
+        """
+        if len(self._training_problems) == 0:
+            logger.debug("No training problem is registered.")
+            return []
+        combined_problems: List[Inference] = []
+        for problem in self._training_problems:
+            logger.debug(f"Observing problem: {problem}")
+            ic = InferenceChain([problem], self._observers)
+            observed_problems = ic.infer()
+            combined_problems = combine_inferences(
+                combined_problems, observed_problems
+            )
+        return combined_problems
+
+    def resolve_problems(self, problems: List[Inference]) -> DiagnosisAction:
+        """
+        Generate the diagnosis action for observed problem
+
+        Args:
+            problems: observed(combined) problems
+
+        Return:
+            diagnosis action
+        """
+        combined_solutions: List[Inference] = []
+        for problem in problems:
+            logger.debug(f"Resolving problem: {problem}")
+            ic = InferenceChain([problem], self._resolvers)
+            input_solutions = ic.infer()
+            if len(input_solutions) > 0:
+                combined_solutions = combine_inferences(
+                    combined_solutions, input_solutions
+                )
+
+        return coordinate_solutions(combined_solutions)
